app/app/dashboard_modules/requirements.txt
streamlit>=1.45.0
google-generativeai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
numpy 